src/squareservice/requirements.txt
grpcio~=1.17
python-json-logger~=0.1.9
protobuf~=3.6
